# Implement faster min/max for string view