# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #6088

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
Better min/max for string view, 2x faster than baselines:

```
string/min nonnull      time:   [133.38 µs 133.41 µs 133.45 µs]
                        thrpt:  [491.10 Melem/s 491.24 Melem/s 491.36 Melem/s]

string/min nullable     time:   [213.53 µs 213.55 µs 213.57 µs]
                        thrpt:  [306.86 Melem/s 306.89 Melem/s 306.92 Melem/s]

string view/min nonnull time:   [106.59 µs 106.68 µs 106.82 µs]
                        thrpt:  [613.54 Melem/s 614.34 Melem/s 614.85 Melem/s]

string view/min nullable
                        time:   [107.06 µs 107.14 µs 107.22 µs]
                        thrpt:  [611.23 Melem/s 611.66 Melem/s 612.13 Melem/s]
```

To reproduce:
```bash
cargo bench --bench aggregate_kernels "string.*/min"
```

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->